pyproject.toml
[tool.poetry]
name = "dtl"
version = "0.1.0"
description = ""
authors = ["Andrew Liu <liu.andrew.x@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.5"
"discord.py" = "^1.3.4"
python-dotenv = "^0.13.0"
humanize = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.3.0"
black = "^19.10b0"
pylint = "^2.5.2"
flake8 = "^3.7.9"
mypy = "^0.770"
jedi = "^0.17.0"

[tool.poetry.scripts]
bot = "dtl.main:entrypoint"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
